01_Assigment-Agent_Run_and_Globle_Levels/agent_level/pyproject.toml
[project]
name = "agent-level"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.0.16",
]
